[Refactor][Linear-Attn] Address review feedback on GLA operators
- Remove unused _gla_bwd_intra_kernel and _gla_bwd_inter_kernel
  (fused kernel subsumes both)
- Rename torch_fwdbwd → ref_autograd_fwdbwd in benchmark for clarity
- Change GLABwdOp default dtype from float32 to float16 for API
  consistency with GLAFwdOp

+++ tileops/ops/gla.py
@@ -130,7 +130,7 @@ class GLABwdOp(Op):
         dim_v: int,
         chunk_size: int = 64,
         scale: float = -1.0,
-        dtype: torch.dtype = torch.float32,
+        dtype: torch.dtype = torch.float16,
         kernel_map: Optional[Dict[str, Kernel]] = None,
         tune: bool = False,
     ) -> None:
